+++ test/tests/model_class_methods.js
@@ -15,7 +15,7 @@ test("all, count, find, first, add, remove", function() {
 
   Post.add(post1, post2).add(post3);
 
-  same(Post.all(), [post1, post2, post3]);
+  same(Post.pluck("id"), [1, 2, 3]);
   equals(Post.count(), 3);
   equals(Post.find(1), post1);
   equals(Post.find(2), post2);
@@ -23,9 +23,9 @@ test("all, count, find, first, add, remove", function() {
   equals(Post.find(4), null);
   equals(Post.first(), post1);
 
-  ok(Post.remove(2));
+  ok(Post.remove(post2));
 
-  same(Post.all(), [post1, post3]);
+  same(Post.pluck("id"), [1, 3]);
   equals(Post.count(), 2);
   equals(Post.find(1), post1);
   ok(Post.find(2) === null);
@@ -37,7 +37,7 @@ test("all, count, find, first, add, remove", function() {
   var post1_duplicate = new Post({ id: 1 });
   Post.add(post1_duplicate);
 
-  same(Post.all(), [post1, post3], "shouldn't be able to add if a model with the same id exists in the collection");
+  same(Post.pluck("id"), [1, 3], "shouldn't be able to add if a model with the same id exists in the collection");
 });
 
 test("detect, select, first, last, count (with chaining)", function() {
@@ -186,14 +186,18 @@ test("sort (and chaining)", function() {
 test("Custom `all` method", function() {
   var Post = Model('post', {
     all: function() {
-      return _.sortBy(this.collection, function(model) {
-        return model.attr("position");
-      });
+      var not_deleted = []
+      var i, model
+      for (i = 0; i < this.collection.length; i++) {
+        model = this.collection[i]
+        if (!model.attr("deleted")) not_deleted.push(model)
+      }
+      return not_deleted
     }
   });
-  var post1 = new Post({ id: 1, position: 2 });
-  var post2 = new Post({ id: 2, position: 3 });
-  var post3 = new Post({ id: 3, position: 1 });
+  var post1 = new Post({ id: 1, deleted: false })
+  var post2 = new Post({ id: 2, deleted: true })
+  var post3 = new Post({ id: 3, deleted: false })
 
   Post.add(post1, post2, post3);
 
@@ -203,7 +207,7 @@ test("Custom `all` method", function() {
     results.push(this.id());
   });
 
-  equals(results.join(", "), "3, 1, 2", "`each` should iterate over `all`");
+  same(results, [1,3], "`each` should iterate over `all`");
 });
 
 test("Custom method with chaining, then more chaining", function() {